+++ frontend/lib/norent/routes.ts
@@ -42,9 +42,7 @@ function createLocalizedRouteInfo(prefix: string) {
       `${prefix}/letter-email-to-user`
     ),
 
-    spanishSurveyEmail: createHtmlEmailStaticPageRouteInfo(
-      `${prefix}/spanish-survey-email`
-    ),
+    oneOffEmail: createHtmlEmailStaticPageRouteInfo(`${prefix}/one-off-email`),
 
     /** The sample letter content (HTML and PDF versions). */
     sampleLetterContent: createLetterStaticPageRouteInfo(
